requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
#psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
dj-database-url
django-filter==24.3
pillow==11.0.0
typing_extensions==4.12.2
sqlparse==0.5.1